Implemented GC-605. Sorted negative eco-news test by classes.
