+++ src/main/java/com/softserve/edu/greencity/api/builders/econews/EcoNewsBuilder.java
@@ -16,7 +16,7 @@ public class EcoNewsBuilder {
     }
 
     public EcoNewsBuilder author(EcoNewsAuthorModel author) {
-        ecoNews.author = author;
+        ecoNews.ecoNewsAuthorDto = author;
         return this;
     }
